# flask-server/rsvp_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "RSVP API"
  version: "1.0.0"
  title: "RSVP API"
  contact:
    email: "aaron@lauterer.at"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/rsvp/api/v1"
schemes:
- "https"
paths:
  /invitation:
    get:
      summary: "get invitation"
      description: "Get an invitation\n"
      operationId: "get_invite"
      parameters:
      - name: "id"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "invite matching id"
          schema:
            $ref: "#/definitions/invitation"
        400:
          description: "bad input parameter"
        404:
          description: "invitation not found"
        500:
          description: "An unexpected error occured."
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rsvp_server.controllers.default_controller"
    post:
      summary: "adds an invitation"
      description: "Adds an invitation"
      operationId: "add_invite"
      parameters:
      - in: "body"
        name: "invitation"
        required: true
        schema:
          $ref: "#/definitions/invitation"
      responses:
        201:
          description: "invitation created"
        400:
          description: "invalid input, object invalid"
        401:
          description: "Access token is missing or invalid"
        409:
          description: "an existing invitation already exists"
        500:
          description: "An unexpected error occured."
          schema:
            $ref: "#/definitions/Error"
      security:
      - bearerAuth: []
      x-swagger-router-controller: "rsvp_server.controllers.default_controller"
    put:
      summary: "update invitation"
      description: "Updates an invitiation."
      operationId: "update_invite"
      parameters:
      - in: "body"
        name: "updateinvitation"
        required: false
        schema:
          $ref: "#/definitions/invitation"
      responses:
        200:
          description: "Invitation updated"
        400:
          description: "invalid input, object invalid"
        404:
          description: "invitation not found"
        500:
          description: "An unexpected error occured."
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rsvp_server.controllers.default_controller"
    delete:
      summary: "deletes an invitation"
      description: "Deletes an invitation"
      operationId: "del_invite"
      parameters:
      - name: "delid"
        in: "query"
        required: true
        type: "string"
        pattern: "^([A-Z0-9]){7}$"
        x-regex: "^([A-Z0-9]){7}$"
        x-modifiers: []
      responses:
        201:
          description: "invitation deleted"
        400:
          description: "invalid input"
        401:
          description: "Access token is missing or invalid"
        404:
          description: "invitation not found"
        500:
          description: "An unexpected error occured."
          schema:
            $ref: "#/definitions/Error"
      security:
      - bearerAuth: []
      x-swagger-router-controller: "rsvp_server.controllers.default_controller"
  /allinvites:
    get:
      summary: "get all invitations"
      description: "Get all invitations"
      operationId: "get_all_invites"
      parameters: []
      responses:
        200:
          description: "Get all invitations"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/invitation"
        401:
          description: "Access token is missing or invalid"
        500:
          description: "An unexpected error occured."
          schema:
            $ref: "#/definitions/Error"
      security:
      - bearerAuth: []
      x-swagger-router-controller: "rsvp_server.controllers.default_controller"
  /login:
    post:
      summary: "login admin user"
      operationId: "login"
      parameters:
      - in: "body"
        name: "login"
        required: false
        schema:
          $ref: "#/definitions/login"
      responses:
        200:
          description: "successful login"
          schema:
            type: "string"
        400:
          description: "login failed"
        500:
          description: "An unexpected error occured."
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rsvp_server.controllers.default_controller"
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  invitation:
    type: "object"
    required:
    - "id"
    - "rsvped"
    properties:
      id:
        type: "string"
        example: "ME4T6ZW"
        pattern: "^([A-Z0-9]){7}$"
        x-regex: "^([A-Z0-9]){7}$"
        x-modifiers: []
      guest:
        type: "array"
        items:
          type: "string"
          example: "Max Muster"
      rsvped:
        type: "boolean"
      date:
        type: "string"
        format: "date-time"
      table:
        type: "boolean"
      email:
        type: "string"
        format: "email"
      phone:
        type: "string"
        example: "+43 660 1234567"
      comment:
        type: "string"
        example: "This is a comment"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      phone: "+43 660 1234567"
      guest:
      - "Max Muster"
      - "Max Muster"
      comment: "This is a comment"
      id: "ME4T6ZW"
      rsvped: true
      table: true
      email: "email"
  Error:
    properties:
      code:
        type: "string"
      message:
        type: "string"
  login:
    type: "object"
    required:
    - "pass"
    - "user"
    properties:
      user:
        type: "string"
      pass:
        type: "string"
        format: "password"
responses:
  Standard500ErrorResponse:
    description: "An unexpected error occured."
    schema:
      $ref: "#/definitions/Error"
  UnauthorizedError:
    description: "Access token is missing or invalid"